requirements.txt
# Core Dependencies - Tested and Working
streamlit==1.47.1
google-generativeai==0.8.5
python-dotenv==1.1.1

# Security & Encryption
cryptography==45.0.5

# Data Processing & Analytics
pandas==2.3.1
numpy==2.3.2

# Visualization
plotly==6.2.0

# NLP & Language Processing
textblob==0.19.0
langdetect==1.0.9

# Additional Utilities
requests==2.32.4
